backend/pyproject.toml
[tool.poetry]
name = "sciencefeed"
version = "1.0.0"
description = "Keep track of your science field using RSS feeds"
authors = ["Crossoufire <vince.delmas.38@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-cors = "^5.0.0"
flask-mail = "^0.10.0"
flask-httpauth = "^4.8.0"
flask-migrate = "^4.0.7"
pillow = "^10.4.0"
pyjwt = "^2.9.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
marshmallow = "^3.22.0"
flask-marshmallow = "^1.2.1"
marshmallow-sqlalchemy = "^1.1.0"
webargs = "^8.4.0"
feedparser = "^6.0.11"
tqdm = "^4.66.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
